requirements.txt
fastapi
uvicorn
python-dotenv
supabase
requests
sentence-transformers
python-multipart
